sql: move sql proto defs to sqlbase
Only refactoring the protos and associated code in this change.
