+++ sql/sqlbase/privilege.proto
@@ -15,8 +15,8 @@
 // Author: Marc Berhault (marc@cockroachlabs.com)
 
 syntax = "proto2";
-package cockroach.sql;
-option go_package = "sql";
+package cockroach.sqlbase;
+option go_package = "sqlbase";
 
 import weak "gogoproto/gogo.proto";
