+++ sql/sqlbase/structured.proto
@@ -16,11 +16,11 @@
 
 // Cannot be proto3 because we use nullable primitives.
 syntax = "proto2";
-package cockroach.sql;
-option go_package = "sql";
+package cockroach.sqlbase;
+option go_package = "sqlbase";
 
 import "cockroach/roachpb/data.proto";
-import "cockroach/sql/privilege.proto";
+import "cockroach/sql/sqlbase/privilege.proto";
 import weak "gogoproto/gogo.proto";
 
 message ColumnType {
